requirements.txt
requests==2.24.0
Flask==1.1.2
python-dotenv==0.15.0
slackclient==2.9.3
